+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorArrayTest.kt
@@ -178,8 +178,7 @@ public class TestArray {
         @Override
         public int hashCode() {
             int hash = id.hashCode();
-            hash ^= name.hashCode();
-            return hash;
+            return hash ^ name.hashCode();
         }
 
     }