+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorCustomClassTest.kt
@@ -158,8 +158,7 @@ public class Person {
     @Override
     public int hashCode() {
         int hash = id.hashCode();
-        hash ^= name.hashCode();
-        return hash;
+        return hash ^ name.hashCode();
     }
 
 }
@@ -219,8 +218,7 @@ public class TestCustom {
     @Override
     public int hashCode() {
         int hash = aaa.hashCode();
-        hash ^= bbb != null ? bbb.hashCode() : 0;
-        return hash;
+        return hash ^ (bbb != null ? bbb.hashCode() : 0);
     }
 
 }