+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorPersonSchemaTest.kt
@@ -154,9 +154,8 @@ public class TestPerson {
     @Override
     public int hashCode() {
         int hash = name.hashCode();
-        hash ^= nickname != null ? nickname.hashCode() : 0;
-        hash ^= age;
-        return hash;
+        hash ^= (nickname != null ? nickname.hashCode() : 0);
+        return hash ^ age;
     }
 
 }